app/grade_extractor.py
from google import genai
import os
from dotenv import load_dotenv

load_dotenv()
client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

def extract_grade(text):
    prompt = f"""
Extract grade level from this sentence.
Return only grade.
If none, return None.

Sentence: {text}
"""

    response = client.models.generate_content(
        model="gemini-2.5-flash",
        contents=prompt
    )

    return response.text.strip()
